Admin notification, automation test

+++ src/main/java/com/backend/dream/rest/ProductRestController.java
@@ -1,16 +1,22 @@
 package com.backend.dream.rest;
 
+import com.backend.dream.dto.NotificationDTO;
 import com.backend.dream.dto.ProductDTO;
 import com.backend.dream.dto.ProductSizeDTO;
 import com.backend.dream.entity.Product;
+import com.backend.dream.service.AccountService;
+import com.backend.dream.service.NotificationService;
 import com.backend.dream.service.ProductService;
 import com.backend.dream.service.ProductSizeService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.sql.Timestamp;
+import java.time.Instant;
 import java.util.List;
 
 @CrossOrigin("*")
@@ -21,6 +27,12 @@ public class ProductRestController {
     private ProductService productService;
     @Autowired
     private ProductSizeService productSizeService;
+
+    @Autowired
+    private AccountService accountService;
+
+    @Autowired
+    private NotificationService notificationService;
 
     @GetMapping("/{id}")
     public ProductDTO getOne(@PathVariable("id") Long id) {
@@ -37,19 +49,83 @@ public class ProductRestController {
     }
 
     @PostMapping()
-    public Product create(@RequestBody ProductDTO productDTO) {
-        return productService.create(productDTO);
+    public Product create(@RequestBody ProductDTO productDTO, HttpServletRequest request) {
+        String username = request.getRemoteUser();
+        Long idAccount = accountService.findIDByUsername(username);
+
+        Long idRole = accountService.findRoleIdByUsername(username);
+        if(idRole == 1 || idRole == 2){
+            String notificationTitle = "Có sự thay đổi trong sản phẩm";
+            String notificationText = "Sản phẩm '" + productDTO.getName() + "' đã được thêm bởi '" + username +"'";
+            NotificationDTO notificationDTO = new NotificationDTO();
+            notificationDTO.setIdAccount(idAccount);
+            notificationDTO.setNotificationTitle(notificationTitle);
+            notificationDTO.setNotificationText(notificationText);
+            notificationDTO.setId_role(idRole);
+            notificationDTO.setImage("product-change.jpg");
+            notificationDTO.setCreatedTime(Timestamp.from(Instant.now()));
+            notificationService.createNotification(notificationDTO);
+            return productService.create(productDTO);
+        }
+        return  null;
     }
 
     @PutMapping("{id}")
-    public ProductDTO update(@RequestBody ProductDTO productDTO, @PathVariable("id") Long id) {
-        return productService.update(productDTO);
+    public ProductDTO update(@RequestBody ProductDTO productDTO, @PathVariable("id") Long id, HttpServletRequest request) {
+        String username = request.getRemoteUser();
+        Long idAccount = accountService.findIDByUsername(username);
+        Long idRole = accountService.findRoleIdByUsername(username);
+
+        if (idRole == 1 || idRole == 2) {
+            ProductDTO previousProduct = productService.findById(id);
+
+            ProductDTO updatedProduct = productService.update(productDTO);
+
+            String notificationTitle = "Có sự thay đổi trong sản phẩm";
+            String notificationText = "Sản phẩm '" + previousProduct.getName() + "' đã được cập nhật bởi '" + username + "'";
+
+            NotificationDTO notificationDTO = new NotificationDTO();
+            notificationDTO.setIdAccount(idAccount);
+            notificationDTO.setNotificationTitle(notificationTitle);
+            notificationDTO.setNotificationText(notificationText);
+            notificationDTO.setId_role(idRole);
+            notificationDTO.setImage("product-change.jpg");
+            notificationDTO.setCreatedTime(Timestamp.from(Instant.now()));
+
+            notificationService.createNotification(notificationDTO);
+
+            return updatedProduct;
+        }
+
+        return null;
     }
 
+
     @DeleteMapping("{id}")
-    public void delete(@PathVariable("id") Long id) {
-        productService.delete(id);
+    public void delete(@PathVariable("id") Long id, HttpServletRequest request) {
+        String username = request.getRemoteUser();
+        Long idAccount = accountService.findIDByUsername(username);
+        Long idRole = accountService.findRoleIdByUsername(username);
+
+        if (idRole == 1 || idRole == 2) {
+            ProductDTO deletedProduct = productService.findById(id);
+
+            String notificationTitle = "Có sự thay đổi trong sản phẩm";
+            String notificationText = "Sản phẩm '" + deletedProduct.getName() + "' đã được xóa bởi '" + username + "'";
+
+            NotificationDTO notificationDTO = new NotificationDTO();
+            notificationDTO.setIdAccount(idAccount);
+            notificationDTO.setNotificationTitle(notificationTitle);
+            notificationDTO.setNotificationText(notificationText);
+            notificationDTO.setId_role(idRole);
+            notificationDTO.setImage("product-change.jpg");
+            notificationDTO.setCreatedTime(Timestamp.from(Instant.now()));
+            notificationService.createNotification(notificationDTO);
+            productService.delete(id);
+        }
+
     }
+
 
     @GetMapping("/getProductPriceByName")
     public ResponseEntity<Double> getProductPriceByName(
