+++ src/main/java/com/backend/dream/rest/ProductSizeRestController.java
@@ -1,12 +1,19 @@
 package com.backend.dream.rest;
 
 
+import com.backend.dream.dto.NotificationDTO;
+import com.backend.dream.dto.ProductDTO;
 import com.backend.dream.dto.ProductSizeDTO;
 import com.backend.dream.entity.ProductSize;
+import com.backend.dream.service.AccountService;
+import com.backend.dream.service.NotificationService;
 import com.backend.dream.service.ProductSizeService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.sql.Timestamp;
+import java.time.Instant;
 import java.util.List;
 
 @CrossOrigin("*")
@@ -15,20 +22,99 @@ import java.util.List;
 public class ProductSizeRestController {
     @Autowired
     private ProductSizeService productSizeService;
+
+    @Autowired
+    private AccountService accountService;
+
+    @Autowired
+    private NotificationService notificationService;
+
     @GetMapping()
     public List<ProductSizeDTO> getAllSizes() {
         return productSizeService.findAll();
     }
     @PostMapping()
-    public ProductSize productSize(@RequestBody ProductSizeDTO productSizeDTO) {
-        return productSizeService.create(productSizeDTO);
+    public ProductSize create(@RequestBody ProductSizeDTO productSizeDTO, HttpServletRequest request) {
+        String username = request.getRemoteUser();
+        Long idAccount = accountService.findIDByUsername(username);
+        Long idRole = accountService.findRoleIdByUsername(username);
+
+        if (idRole == 1 || idRole == 2) {
+            ProductSize createdProductSize = productSizeService.create(productSizeDTO);
+
+            String productName = productSizeService.getProductNameByProductSizeID(productSizeDTO.getId_product());
+
+            String notificationTitle = "Có sự thay đổi giá sản phẩm";
+            String notificationText = "Giá của sản phẩm '" + productName + "' đã được thêm bởi '" + username + "'";
+
+            NotificationDTO notificationDTO = new NotificationDTO();
+            notificationDTO.setIdAccount(idAccount);
+            notificationDTO.setNotificationTitle(notificationTitle);
+            notificationDTO.setNotificationText(notificationText);
+            notificationDTO.setId_role(idRole);
+            notificationDTO.setImage("size-change.jpg");
+            notificationDTO.setCreatedTime(Timestamp.from(Instant.now()));
+            notificationService.createNotification(notificationDTO);
+
+            return createdProductSize;
+        }
+
+        return null;
     }
-    @DeleteMapping("{id}")
-    public void delete(@PathVariable("id") Long id) {
-        productSizeService.delete(id);
-    }
+
+
     @PutMapping("{id}")
-    public ProductSizeDTO productSizeDTO(@RequestBody ProductSizeDTO productSizeDTO,@PathVariable("id") Long id){
-        return productSizeService.update(productSizeDTO,id);
+    public ProductSizeDTO update(@RequestBody ProductSizeDTO productSizeDTO, @PathVariable("id") Long id, HttpServletRequest request) {
+        String username = request.getRemoteUser();
+        Long idAccount = accountService.findIDByUsername(username);
+        Long idRole = accountService.findRoleIdByUsername(username);
+
+        if (idRole == 1 || idRole == 2) {
+            ProductSizeDTO previousProductSize = productSizeService.findByID(id);
+            ProductSizeDTO updatedProductSize = productSizeService.update(productSizeDTO, id);
+
+            String productName = productSizeService.getProductNameByProductSizeID(id);
+            String notificationTitle = "Có sự thay đổi giá sản phẩm";
+            String notificationText = "Giá của sản phẩm '" + productName + "' đã được cập nhật bởi '" + username + "'";
+
+            NotificationDTO notificationDTO = new NotificationDTO();
+            notificationDTO.setIdAccount(idAccount);
+            notificationDTO.setNotificationTitle(notificationTitle);
+            notificationDTO.setNotificationText(notificationText);
+            notificationDTO.setId_role(idRole);
+            notificationDTO.setImage("size-change.jpg");
+            notificationDTO.setCreatedTime(Timestamp.from(Instant.now()));
+            notificationService.createNotification(notificationDTO);
+
+            return updatedProductSize;
+        }
+
+        return null;
     }
+
+    @DeleteMapping("{id}")
+    public void delete(@PathVariable("id") Long id, HttpServletRequest request) {
+        String username = request.getRemoteUser();
+        Long idAccount = accountService.findIDByUsername(username);
+        Long idRole = accountService.findRoleIdByUsername(username);
+
+        if (idRole == 1 || idRole == 2) {
+            String productName = productSizeService.getProductNameByProductSizeID(id);
+            if (productName != null) {
+                String notificationTitle = "Có sự thay đổi trong sản phẩm";
+                String notificationText = "Một giá của sản phẩm '" + productName + "' đã được xóa bởi '" + username + "'";
+
+                NotificationDTO notificationDTO = new NotificationDTO();
+                notificationDTO.setIdAccount(idAccount);
+                notificationDTO.setNotificationTitle(notificationTitle);
+                notificationDTO.setNotificationText(notificationText);
+                notificationDTO.setId_role(idRole);
+                notificationDTO.setImage("size-change.jpg");
+                notificationDTO.setCreatedTime(Timestamp.from(Instant.now()));
+                notificationService.createNotification(notificationDTO);
+                productSizeService.delete(id);
+            }
+        }
+    }
+
 }
